pyproject.toml
[project]
name = "astragalaxy"
version = "0.1.0"
description = "API-based space game"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "dishka>=1.6.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.8",
]
